+++ app/src/main/java/io/github/sds100/keymapper/mappings/keymaps/KeyMapListViewModel.kt
@@ -539,7 +539,9 @@ class KeyMapListViewModel(
     }
 
     fun onEnableGuiKeyboardClick() {
-        setupGuiKeyboard.enableInputMethod()
+        coroutineScope.launch {
+            setupGuiKeyboard.enableInputMethod()
+        }
     }
 
     fun onChooseGuiKeyboardClick() {