+++ app/src/main/java/io/github/sds100/keymapper/system/BootBroadcastReceiver.kt
@@ -3,7 +3,7 @@ package io.github.sds100.keymapper.system
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import io.github.sds100.keymapper.ServiceLocator
+import io.github.sds100.keymapper.KeyMapperApp
 
 /**
  * Created by sds100 on 24/03/2019.
@@ -12,12 +12,9 @@ import io.github.sds100.keymapper.ServiceLocator
 class BootBroadcastReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context?, intent: Intent?) {
         context ?: return
-        if (intent?.action == Intent.ACTION_BOOT_COMPLETED) {
-            /*
-            Initializing the controller will update any notifications since it will collect the values
-            in the constructor
-             */
-            ServiceLocator.notificationController(context)
+
+        if (intent?.action == Intent.ACTION_LOCKED_BOOT_COMPLETED) {
+            (context.applicationContext as? KeyMapperApp)?.onBootUnlocked()
         }
     }
 }