+++ app/src/main/java/io/github/sds100/keymapper/system/accessibility/BaseAccessibilityServiceController.kt
@@ -150,6 +150,10 @@ abstract class BaseAccessibilityServiceController(
             flags = flags.withFlag(AccessibilityServiceInfo.FLAG_ENABLE_ACCESSIBILITY_VOLUME)
         }
 
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            flags = flags.withFlag(AccessibilityServiceInfo.FLAG_INPUT_METHOD_EDITOR)
+        }
+
         return@lazy flags
     }
 
@@ -505,6 +509,11 @@ abstract class BaseAccessibilityServiceController(
             }
 
             is ServiceEvent.TriggerKeyMap -> triggerKeyMapFromIntent(event.uid)
+
+            is ServiceEvent.EnableInputMethod -> if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+                accessibilityService.setInputMethodEnabled(event.imeId, true)
+            }
+
             else -> Unit
         }
     }