+++ app/src/main/java/io/github/sds100/keymapper/system/accessibility/MyAccessibilityService.kt
@@ -4,6 +4,7 @@ import android.accessibilityservice.AccessibilityService
 import android.accessibilityservice.FingerprintGestureController
 import android.accessibilityservice.GestureDescription
 import android.accessibilityservice.GestureDescription.StrokeDescription
+import android.annotation.SuppressLint
 import android.app.ActivityManager
 import android.content.Intent
 import android.content.res.Configuration
@@ -560,4 +561,17 @@ class MyAccessibilityService :
     }
 
     override fun findFocussedNode(focus: Int): AccessibilityNodeModel? = findFocus(focus)?.toModel()
+
+    override fun inputText(text: String) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            inputMethod?.currentInputConnection?.commitText(text, 1, null)
+        }
+    }
+
+    override fun setInputMethodEnabled(imeId: String, enabled: Boolean) {
+        @SuppressLint("CheckResult")
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            softKeyboardController.setInputMethodEnabled(imeId, enabled)
+        }
+    }
 }