+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## [3.0 Beta 4](https://github.com/sds100/KeyMapper/releases/tag/v3.0.0-beta.4)
+
+#### 2 April 2025
+
+_See the changes from previous 3.0 Beta releases as well._
+
+## Added
+
+- #1620 enable Key Mapper Basic Input Method without user interaction on Android 13+.
+- #1619 Automatically select the non key mapper keyboard when the device is locked and wanting to type.
+
+## Changed
+
+- *Finally* renamed the theme settings after many years. @jambl3r.
+
+## Bug fixes
+
+- #1618, #1532, #1590 The Key Mapper keyboard is no longer required for Text actions.
+- Flashlight action works again on devices that do not support variable brightness
+
 ## [3.0 Beta 3](https://github.com/sds100/KeyMapper/releases/tag/v3.0.0-beta.3)
 
 _See the changes from previous 3.0 Beta releases as well._